Fixing a horrible task_exit bug
Took all day to track down a bug that was corrupting the stack
for exiting threads. Turns out I'm an idiot and was freeing the stack
for a task while it was still in use. New plan here is to always
have some task running for each cpu that cleans up deleted tasks
from a separate task.